+++ R/ggplot_boxplot.R
@@ -118,7 +118,7 @@ theme_boxplot <-
 #' @param stat String of "boxplot" or "identity". Defaults to "boxplot". If identity is selected, data provided must be grouped by the x_var with ymin, lower, middle, upper, ymax variables. Note "identity" does not provide outliers.
 #' @param pal Character vector of hex codes. Defaults to viridis. Use the pals package to find a suitable palette.
 #' @param width Width of the box. Defaults to 0.5.
-#' @param fill_alpha The alpha of the fill. Defaults to 0.1. 
+#' @param alpha The alpha of the fill. Defaults to 0.1. 
 #' @param point_size The size of the outliers. Defaults to 1.
 #' @param line_size The size of the outlines of boxplots. Defaults to 0.5.
 #' @param title Title string. Defaults to "[Title]".
@@ -174,7 +174,7 @@ ggplot_boxplot <- function(data,
                        stat = "boxplot",
                        pal = NULL,
                        width = 0.5,
-                       fill_alpha = 0.1,
+                       alpha = 0.1,
                        line_size = 0.5,
                        point_size = 1, 
                        title = "[Title]",
@@ -255,7 +255,7 @@ ggplot_boxplot <- function(data,
           fill = pal,
           width = width,
           size = line_size, 
-          alpha = fill_alpha,
+          alpha = alpha,
           outlier.alpha = 1, 
           outlier.size = point_size 
         )
@@ -269,7 +269,7 @@ ggplot_boxplot <- function(data,
           fill = pal,
           width = width,
           size = line_size, 
-          alpha = fill_alpha,
+          alpha = alpha,
           outlier.alpha = 1, 
           outlier.size = point_size
         )
@@ -291,7 +291,7 @@ ggplot_boxplot <- function(data,
         fill = pal,
         width = width,
         size = line_size, 
-        alpha = fill_alpha,
+        alpha = alpha,
         outlier.alpha = 1, 
         outlier.size = point_size
       )
@@ -430,7 +430,7 @@ ggplot_boxplot <- function(data,
 #' @param stat String of "boxplot" or "identity". Defaults to "boxplot". If identity is selected, data provided must be grouped by the x_var and facet_var with ymin, lower, middle, upper, ymax variables. Note "identity" does not provide outliers.
 #' @param pal Character vector of hex codes. Defaults to viridis. Use the pals package to find a suitable palette.
 #' @param width Width of the box. Defaults to 0.5.
-#' @param fill_alpha The alpha of the fill. Defaults to 0.1. 
+#' @param alpha The alpha of the fill. Defaults to 0.1. 
 #' @param line_size The size of the outlines of boxplots. Defaults to 0.5.
 #' @param point_size The size of the outliers. Defaults to 1.
 #' @param title Title string. Defaults to "[Title]".
@@ -479,7 +479,7 @@ ggplot_boxplot_facet <-
            stat = "boxplot",
            pal = NULL,
            width = 0.5,
-           fill_alpha = 0.1,
+           alpha = 0.1,
            line_size = 0.5,
            point_size = 1,
            title = "[Title]",
@@ -559,7 +559,7 @@ ggplot_boxplot_facet <-
             fill = pal,
             width = width,
             size = line_size, 
-            alpha = fill_alpha,
+            alpha = alpha,
             outlier.alpha = 1
           )
       }
@@ -572,7 +572,7 @@ ggplot_boxplot_facet <-
             fill = pal,
             width = width,
             size = line_size, 
-            alpha = fill_alpha,
+            alpha = alpha,
             outlier.alpha = 1, 
             outlier.size = point_size
           )
@@ -600,7 +600,7 @@ ggplot_boxplot_facet <-
           fill = pal,
           width = width,
           size = line_size, 
-          alpha = fill_alpha,
+          alpha = alpha,
           outlier.alpha = 1, 
           outlier.size = point_size
         )