+++ R/ggplot_point.R
@@ -159,7 +159,7 @@ ggplot_point <- function(data,
                          x_var,
                          y_var,
                          text_var = NULL,
-                         point_size = 1.5,
+                         point_size = 1,
                          pal = NULL, 
                          title = "[Title]",
                          title_wrap = 70,
@@ -368,7 +368,7 @@ ggplot_point <- function(data,
 #' @param y_zero TRUE or FALSE whether the minimum of the y scale is zero. Defaults to TRUE.
 #' @param y_zero_line TRUE or FALSE whether to add a zero reference line to the y axis. Defaults to NULL, which is TRUE if there are positive and negative values in y_var. Otherwise it is FALSE.
 #' @param col_labels_dp Select the appropriate number of decimal places for numeric variable auto legend labels. Defaults to 1.
-#' @param col_labels Adjust the  x scale labels through a vector.
+#' @param col_labels Adjust the colour scale labels through a vector.
 #' @param col_labels_ncol The number of columns in the legend. Defaults to 1.
 #' @param col_labels_nrow The number of rows in the legend.
 #' @param col_title Colour title string for the legend. Defaults to NULL.
@@ -394,7 +394,7 @@ ggplot_point_col <-
            y_var,
            col_var,
            text_var = NULL,
-           point_size = 1.5,
+           point_size = 1,
            pal = NULL,
            pal_rev = FALSE,
            col_method = NULL,
@@ -698,7 +698,7 @@ ggplot_point_facet <-
            y_var,
            facet_var,
            text_var = NULL,
-           point_size = 1.5,
+           point_size = 1,
            pal = NULL,
            title = "[Title]",
            title_wrap = 70,
@@ -907,7 +907,7 @@ ggplot_point_facet <-
 #' @param col_labels_dp Select the appropriate number of decimal places for numeric variable auto legend labels. Defaults to 1.
 #' @param col_labels_ncol The number of columns in the legend. Defaults to 1.
 #' @param col_labels_nrow The number of rows in the legend.
-#' @param col_labels Adjust the  x scale labels through a vector.
+#' @param col_labels Adjust the colour scale labels through a vector.
 #' @param col_method The method of colouring features, either "bin", "quantile" or "category." If numeric, defaults to "quantile".
 #' @param col_na TRUE or FALSE of whether to show NA values of the colour variable.
 #' @param col_quantile_by_facet TRUE of FALSE whether quantiles should be calculated for each group of the facet variable. Defaults to TRUE.
@@ -939,7 +939,7 @@ ggplot_point_col_facet <-
            col_var,
            facet_var,
            text_var = NULL,
-           point_size = 1.5,
+           point_size = 1,
            pal = NULL,
            pal_rev = FALSE,
            title = "[Title]",