+++ R/leaflet_sf.R
@@ -5,10 +5,9 @@
 #' @param data An sf object of geometry type point/multipoint, linestring/multilinestring or polygon/multipolygon geometry type. Required input.
 #' @param popup_var Quoted variable of a variable to include in the popup. If NULL, defaults to making a leafpop::popupTable of all columns.
 #' @param pal Character vector of hex codes. Defaults to viridis. Use the pals package to find a suitable palette.
-#' @param point_size Size of points. Defaults to 1.
-#' @param line_size Size of lines around features. Defaults to 2.
-#' @param fill_alpha The opacity of the fill within features. Defaults to 0.1. 
-#' @param line_alpha The opacity of the lines around features. Defaults to 1. 
+#' @param point_size Size of points (i.e. radius). Defaults to 1.
+#' @param line_size Size of lines around features (i.e. weight). Defaults to 2.
+#' @param alpha The opacity of the fill within features (i.e. fillOpacity). Defaults to 0.1. 
 #' @param basemap The underlying basemap. Either "light", "dark", "satellite", "street", or "ocean". Defaults to "light". Only applicable where shiny equals FALSE.
 #' @param title A title string that will be wrapped into the legend. Defaults to "Title"
 #' @param col_labels_dp Select the appropriate number of decimal places for numeric variable auto legend labels. Defaults to 1.
@@ -23,8 +22,7 @@ leaflet_sf <- function(data,
                        pal = NULL,
                        point_size = 1,
                        line_size = 2,
-                       fill_alpha = 0.1,
-                       line_alpha = 1,
+                       alpha = 0.1,
                        basemap = "light",
                        title = "[Title]",
                        col_labels_dp = 1,
@@ -78,8 +76,8 @@ leaflet_sf <- function(data,
           popup = ~ popup,
           color = pal[1],
           radius = point_size,
-          fillOpacity = fill_alpha,
-          opacity = line_alpha,
+          fillOpacity = alpha,
+          opacity = 1,
           weight = line_size
         ) 
     }
@@ -93,8 +91,8 @@ leaflet_sf <- function(data,
           popup = ~ popup,
           color = pal[1],
           radius = point_size,
-          fillOpacity = fill_alpha,
-          opacity = line_alpha,
+          fillOpacity = alpha,
+          opacity = 1,
           weight = line_size
         ) 
     }
@@ -106,7 +104,7 @@ leaflet_sf <- function(data,
         labels =  col_labels, 
         title = stringr::str_replace_all(stringr::str_wrap(title, 20), "\n", "</br>"),
         position = "bottomright",
-        opacity = line_alpha,
+        opacity = 1,
         labFormat = labelFormat(between = "&ndash;", digits =  col_labels_dp)
       )
   }
@@ -120,8 +118,8 @@ leaflet_sf <- function(data,
           data = data,
           popup = ~ popup,
           color = pal[1],
-          fillOpacity = fill_alpha,
-          opacity = line_alpha,
+          fillOpacity = alpha,
+          opacity = 1,
           weight = line_size
         ) 
     }
@@ -134,8 +132,8 @@ leaflet_sf <- function(data,
           data = data,
           popup = ~ popup,
           color = pal[1],
-          fillOpacity = fill_alpha,
-          opacity = line_alpha,
+          fillOpacity = alpha,
+          opacity = 1,
           weight = line_size
         ) 
     }
@@ -147,7 +145,7 @@ leaflet_sf <- function(data,
         labels =  col_labels, 
         title = stringr::str_replace_all(stringr::str_wrap(title, 20), "\n", "</br>"),
         position = "bottomright",
-        opacity = line_alpha,
+        opacity = 1,
         labFormat = labelFormat(between = "&ndash;", digits =  col_labels_dp)
       )
   }
@@ -161,8 +159,8 @@ leaflet_sf <- function(data,
           data = data,
           popup = ~ popup,
           color = pal[1],
-          fillOpacity = fill_alpha, 
-          opacity = line_alpha,
+          fillOpacity = alpha, 
+          opacity = 1,
           weight = line_size
         ) 
     }
@@ -174,8 +172,8 @@ leaflet_sf <- function(data,
           data = data,
           popup = ~ popup,
           color = pal[1],
-          fillOpacity = fill_alpha, 
-          opacity = line_alpha,
+          fillOpacity = alpha, 
+          opacity = 1,
           weight = line_size
         )
     }
@@ -187,7 +185,7 @@ leaflet_sf <- function(data,
         labels =  col_labels, 
         title = stringr::str_replace_all(stringr::str_wrap(title, 20), "\n", "</br>"),
         position = "bottomright",
-        opacity = line_alpha,
+        opacity = 1,
         labFormat = labelFormat(between = "&ndash;", digits =  col_labels_dp)
       )
     
@@ -203,15 +201,13 @@ leaflet_sf <- function(data,
 #' @param popup_var Quoted variable of a variable to include in the popup. If NULL, defaults to making a leafpop::popupTable of all columns.
 #' @param pal Character vector of hex codes. Defaults to viridis. Use the pals package to find a suitable palette.
 #' @param pal_rev Reverses the palette. Defaults to FALSE.
-#' @param point_size Size of points. Defaults to 1.
-#' @param line_size Size of lines around features. Defaults to 2.
-#' @param fill_alpha The opacity of the fill within features. Defaults to 0.1. 
-#' @param line_alpha The opacity of the lines around features. Defaults to 1. 
+#' @param point_size Size of points (i.e. radius). Defaults to 1.
+#' @param line_size Size of lines around features (i.e. weight). Defaults to 2.
+#' @param alpha The opacity of the fill within features (i.e. fillOpacity). Defaults to 0.1. 
 #' @param basemap The underlying basemap. Either "light", "dark", "satellite", "street", or "ocean". Defaults to "light". Only applicable where shiny equals FALSE.
 #' @param title A title string that will be wrapped into the legend. Defaults to "Title".
 #' @param col_cuts A vector of cuts to colour a numeric variable. If "bin" is selected, the first number in the vector should be either -Inf or 0, and the final number Inf. If "quantile" is selected, the first number in the vector should be 0 and the final number should be 1. Defaults to quartiles. 
 #' @param col_labels_dp Select the appropriate number of decimal places for numeric variable auto legend labels. Defaults to 1.
-#' @param col_drop TRUE or FALSE of whether to drop unused levels from the legend. Defaults to FALSE.
 #' @param col_labels Adjust the  x scale labels through a vector.
 #' @param col_method The method of colouring features, either "bin", "quantile" or "category." if categorical colour variable, NULL results in "category". If numeric variable, defaults to "quantile". Note all numeric variables are cut to be inclusive of the min in the range, and exclusive of the max in the range (except for the final bucket which includes the highest value).
 #' @param col_na TRUE or FALSE  of whether to include NAs of the colour variable. Defaults to TRUE.
@@ -239,13 +235,11 @@ leaflet_sf_col <- function(data,
                            pal_rev = FALSE,
                            point_size = 1,
                            line_size = 2,
-                           fill_alpha = 0.9,
-                           line_alpha = 1,
+                           alpha = 0.9,
                            basemap = "light",
                            title = "[Title]",
                            col_cuts = NULL,
                            col_labels_dp = 1,
-                           col_drop = FALSE,
                            col_labels = NULL,
                            col_method = NULL,
                            col_na = TRUE,
@@ -274,8 +268,8 @@ leaflet_sf_col <- function(data,
   
   if (col_method == "category") {
     if (is.null( col_labels)){ 
-      if (is.factor(col_var_vctr) &  col_drop == FALSE) labels <- levels(col_var_vctr)
-      else if (is.character(col_var_vctr) | col_drop == TRUE) labels <- sort(unique(col_var_vctr))
+      if (is.factor(col_var_vctr)) labels <- levels(col_var_vctr)
+      else if (is.character(col_var_vctr)) labels <- sort(unique(col_var_vctr))
     }
     else if (!is.null( col_labels)) labels <-  col_labels
     
@@ -377,8 +371,8 @@ leaflet_sf_col <- function(data,
           label = ~ htmltools::htmlEscape(text_var_vctr),
           popup = ~ popup,
           radius = point_size,
-          fillOpacity = fill_alpha,
-          opacity = line_alpha,
+          fillOpacity = alpha,
+          opacity = 1,
           weight = line_size
         )
     }
@@ -392,8 +386,8 @@ leaflet_sf_col <- function(data,
           label = ~ htmltools::htmlEscape(text_var_vctr),
           popup = ~ popup,
           radius = point_size,
-          fillOpacity = fill_alpha,
-          opacity = line_alpha,
+          fillOpacity = alpha,
+          opacity = 1,
           weight = line_size
         ) 
     }
@@ -405,7 +399,7 @@ leaflet_sf_col <- function(data,
         labels = labels,
         title = stringr::str_replace_all(stringr::str_wrap(title, 20), "\n", "</br>"),
         position = "bottomright",
-        opacity = line_alpha,
+        opacity = 1,
         labFormat = labelFormat(between = "&ndash;", digits =  col_labels_dp)
       )
     
@@ -420,8 +414,8 @@ leaflet_sf_col <- function(data,
           color = ~ pal_fun(col_var_vctr),
           popup = ~ popup,
           label = ~ htmltools::htmlEscape(text_var_vctr),
-          fillOpacity = fill_alpha,
-          opacity = line_alpha,
+          fillOpacity = alpha,
+          opacity = 1,
           weight = line_size
         ) 
     }
@@ -434,8 +428,8 @@ leaflet_sf_col <- function(data,
           color = ~ pal_fun(col_var_vctr),
           popup = ~ popup,
           label = ~ htmltools::htmlEscape(text_var_vctr),
-          fillOpacity = fill_alpha,
-          opacity = line_alpha,
+          fillOpacity = alpha,
+          opacity = 1,
           weight = line_size
         ) 
     }
@@ -447,7 +441,7 @@ leaflet_sf_col <- function(data,
         labels = labels,
         title = stringr::str_replace_all(stringr::str_wrap(title, 20), "\n", "</br>"),
         position = "bottomright",
-        opacity = line_alpha,
+        opacity = 1,
         labFormat = labelFormat(between = "&ndash;", digits =  col_labels_dp)
       )
     
@@ -461,8 +455,8 @@ leaflet_sf_col <- function(data,
           color = ~ pal_fun(col_var_vctr),
           popup = ~ popup,
           label = ~ htmltools::htmlEscape(text_var_vctr),
-          fillOpacity = fill_alpha, 
-          opacity = line_alpha,
+          fillOpacity = alpha, 
+          opacity = 1,
           weight = line_size
         ) 
     }
@@ -475,8 +469,8 @@ leaflet_sf_col <- function(data,
           color = ~ pal_fun(col_var_vctr),
           popup = ~ popup,
           label = ~ htmltools::htmlEscape(text_var_vctr),
-          fillOpacity = fill_alpha, 
-          opacity = line_alpha,
+          fillOpacity = alpha, 
+          opacity = 1,
           weight = line_size
         ) 
     }
@@ -488,7 +482,7 @@ leaflet_sf_col <- function(data,
         labels = labels,
         title = stringr::str_replace_all(stringr::str_wrap(title, 20), "\n", "</br>"),
         position = "bottomright",
-        opacity = line_alpha,
+        opacity = 1,
         labFormat = labelFormat(between = "&ndash;", digits =  col_labels_dp)
       )
   }